HOTFIX / refactor gestion des outils twig du CMS dans le ContentRenderer (passage via une TwigExtension)
